refactor(canon): split path helpers and rewrite source map into siblings
