+++ crates/vize_canon/src/corsa_bridge.rs
@@ -9,6 +9,7 @@ mod session;
 mod types;
 mod vue_dependencies;
 mod vue_dependencies_alias;
+mod vue_dependency_paths;
 mod vue_dependency_specifiers;
 mod vue_document;
 #[cfg(test)]
